\name{plethyData}
\docType{data}
\alias{plethyData}
\alias{plethy.example.file}
\title{Experimental Data Package for plethy}
\description{Package containing data accompanying the Journal of Statistical Computing paper}
\usage{
plethy.example.file()
}

\examples{
  plethy.example.file()
}
\keyword{datasets}
